+++ apps/web/modules/saas/settings/components/ChangePassword.tsx
@@ -11,11 +11,11 @@ import {
 	FormMessage,
 } from "@repo/ui/components/form";
 import { PasswordInput } from "@repo/ui/components/password-input";
+import { toastError, toastSuccess } from "@repo/ui/components/toast";
 import { SettingsItem } from "@saas/shared/components/SettingsItem";
 import { useRouter } from "@shared/hooks/router";
 import { useTranslations } from "next-intl";
 import { useForm } from "react-hook-form";
-import { toast } from "sonner";
 import { z } from "zod";
 
 const formSchema = z.object({
@@ -42,7 +42,7 @@ export function ChangePasswordForm() {
 		});
 
 		if (error) {
-			toast.error(
+			toastError(
 				t(
 					"settings.account.security.changePassword.notifications.error",
 				),
@@ -51,7 +51,7 @@ export function ChangePasswordForm() {
 			return;
 		}
 
-		toast.success(
+		toastSuccess(
 			t("settings.account.security.changePassword.notifications.success"),
 		);
 		form.reset({});